+++ packages/api-page-builder-so-ddb-es/src/definitions/pageEntity.ts
@@ -67,15 +67,6 @@ export const createPageEntity = (params: Params): Entity<any> => {
             status: {
                 type: "string"
             },
-            visibility: {
-                type: "map"
-            },
-            home: {
-                type: "boolean"
-            },
-            notFound: {
-                type: "boolean"
-            },
             createdOn: {
                 type: "string"
             },